backend/pyproject.toml
[project]
name = "algo-sensei-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "QuangMinhNguyen27405",email = "quangminh27405@gmail.com"}
]

readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.123.5,<0.124.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "cloud-sql-python-connector[pg8000] (>=1.18.5,<2.0.0)",
    "google-cloud-storage (>=3.6.0,<4.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "google-adk[a2a] (>=1.21.0,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.11.0,<26.0.0)",
    "ruff (>=0.14.7,<0.15.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
